Completed a bit of output
