+++ examples/Makefile
@@ -20,11 +20,11 @@ LIB_DIR     = ../lib
 BIN_DIR     = ../bin
 
 # The next line contains the list of utility objects created by this Makefile.
-EXECUTABLES = covid_example
+EXECUTABLES = politics_example
 
-covid_example: $(INC_DIR)/myutils.h $(INC_DIR)/CellularAutomata.h
-	$(CXX) $(CXXFLAGS) covid_example covid_example.cxx -I$(INC_DIR) $(LIB_DIR)/utils.o $(LIB_DIR)/CA_input.o $(LIB_DIR)/CA_output.o
-	mv covid_example $(BIN_DIR)
+politics_example: $(INC_DIR)/myutils.h $(INC_DIR)/CellularAutomata.h
+	$(CXX) $(CXXFLAGS) politics_example politics_example.cxx -I$(INC_DIR) $(LIB_DIR)/utils.o $(LIB_DIR)/CA_input.o $(LIB_DIR)/CA_output.o
+	mv politics_example $(BIN_DIR)
 
 all: $(EXECUTABLES)
