+++ src/input/CA_input.cxx
@@ -247,28 +247,3 @@ int CellularAutomata::purity_rule(int x, int y, std::vector<int> neighborhood){
 
   return sum % _num_states;
 }
-
-// Temporary print that is being used to test if the Initialize function is working as intended.
-// Print the formatted matrix out to the terminal using std::cout. Each row is printed with the first element of the Cellular Automata following an opening square bracket and all elements being seperated by commas. The last element of the Cellular Automata is also followed by a closing square bracket.
-void CellularAutomata::print()
-{
-    for(int i = 0; i < _rows; i++)
-    {
-        for(int j = 0; j < _columns; j++)
-        {
-            if ((i == _rows-1) && (j == _columns-1))
-            {
-                std::cout << _data[i][j] << "]";
-            }
-            else if ((i == 0) && (j == 0))
-            {
-                std::cout << "[" << _data[i][j] << ", ";
-            }
-            else
-            {
-                std::cout << _data[i][j] << ", ";
-            }
-        }
-        std::cout << std::endl;
-    }
-}
